tests: Move code from `utils` to `clientcmd`
The `utils` file is overloaded with a large amount of functionality
in various subjects.
It makes it challenging to handle dependencies and organize the codebase
in a maintainable manner.

This change extracts the functions that handle terminal commands
to a separate sub-package.
