+++ tests/monitoring/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     deps = [
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests:go_default_library",
+        "//tests/clientcmd:go_default_library",
         "//tests/flags:go_default_library",
         "//tests/framework/checks:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
