+++ tests/template_test.go
@@ -27,6 +27,7 @@ import (
 	"strings"
 	"time"
 
+	"kubevirt.io/kubevirt/tests/clientcmd"
 	"kubevirt.io/kubevirt/tests/framework/checks"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -65,7 +66,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 		virtClient, err = kubecli.GetKubevirtClient()
 		util.PanicOnError(err)
 
-		tests.SkipIfNoCmd("oc")
+		clientcmd.SkipIfNoCmd("oc")
 		tests.BeforeTestCleanup()
 		SetDefaultEventuallyTimeout(120 * time.Second)
 		SetDefaultEventuallyPollingInterval(2 * time.Second)
@@ -148,7 +149,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 				for param, value := range templateParams {
 					ocProcessCommand = append(ocProcessCommand, "-p", fmt.Sprintf("%s=%s", param, value))
 				}
-				out, stderr, err := tests.RunCommandPipe(ocProcessCommand, []string{"oc", "create", "-f", "-"})
+				out, stderr, err := clientcmd.RunCommandPipe(ocProcessCommand, []string{"oc", "create", "-f", "-"})
 				ExpectWithOffset(1, err).ToNot(HaveOccurred(), "failed to create VirtualMachine %q via command \"%s | oc create -f -\": %s: %v", vmName, strings.Join(ocProcessCommand, " "), out+stderr, err)
 				ExpectWithOffset(1, out).To(MatchRegexp(`"?%s"? created\n`, vmName), "command \"%s | oc create -f -\" did not print expected message: %s", strings.Join(ocProcessCommand, " "), out+stderr)
 				By("Checking if the VirtualMachine exists")
@@ -166,7 +167,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 				for param, value := range templateParams {
 					ocProcessCommand = append(ocProcessCommand, "-p", fmt.Sprintf("%s=%s", param, value))
 				}
-				out, stderr, err := tests.RunCommandPipe(ocProcessCommand, []string{"oc", "create", "-f", "-"})
+				out, stderr, err := clientcmd.RunCommandPipe(ocProcessCommand, []string{"oc", "create", "-f", "-"})
 				ExpectWithOffset(1, err).To(HaveOccurred(), "creation of VirtualMachine %q via command \"%s | oc create -f -\" succeeded: %s: %v", vmName, strings.Join(ocProcessCommand, " "), out+stderr, err)
 			}
 		}
@@ -174,7 +175,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 		AssertVMDeletionSuccess := func() func() {
 			return func() {
 				By("Deleting the VirtualMachine via oc command")
-				out, stderr, err := tests.RunCommand("oc", "delete", "vm", vmName)
+				out, stderr, err := clientcmd.RunCommand("oc", "delete", "vm", vmName)
 				ExpectWithOffset(1, err).ToNot(HaveOccurred(), "failed to delete VirtualMachine via command \"oc delete vm %s\": %s: %v", vmName, out+stderr, err)
 				ExpectWithOffset(1, out).To(MatchRegexp(`"?%s"? deleted\n`, vmName), "command \"oc delete vm %s\" did not print expected message: %s", vmName, out)
 
@@ -189,7 +190,7 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 		AssertVMDeletionFailure := func() func() {
 			return func() {
 				By("Deleting the VirtualMachine via oc command")
-				out, stderr, err := tests.RunCommand("oc", "delete", "vm", vmName)
+				out, stderr, err := clientcmd.RunCommand("oc", "delete", "vm", vmName)
 				ExpectWithOffset(1, err).To(HaveOccurred(), "failed to delete VirtualMachine via command \"oc delete vm %s\": %s: %v", vmName, out+stderr, err)
 			}
 		}
@@ -200,13 +201,13 @@ var _ = Describe("[Serial][sig-compute]Templates", func() {
 				case "oc":
 					By("Starting VirtualMachine via oc command")
 					patch := `{"spec":{"running":true}}`
-					out, stderr, err := tests.RunCommand("oc", "patch", "vm", vmName, "--type=merge", "-p", patch)
+					out, stderr, err := clientcmd.RunCommand("oc", "patch", "vm", vmName, "--type=merge", "-p", patch)
 					ExpectWithOffset(1, err).ToNot(HaveOccurred(), "failed schedule VirtualMachine %q start via command \"oc patch vm %s --type=merge -p '%s'\": %s: %v", vmName, vmName, patch, out+stderr, err)
 					ExpectWithOffset(1, out).To(MatchRegexp(`"?%s"? patched\n`, vmName), "command \"oc patch vm %s --type=merge -p '%s'\" did not print expected message: %s", vmName, patch, out+stderr)
 
 				case "virtctl":
 					By("Starting VirtualMachine via virtctl command")
-					out, stderr, err := tests.RunCommand("virtctl", "start", vmName)
+					out, stderr, err := clientcmd.RunCommand("virtctl", "start", vmName)
 					ExpectWithOffset(1, err).ToNot(HaveOccurred(), "failed to schedule VirtualMachine %q start via command \"virtctl start %s\": %s: %v", vmName, vmName, out+stderr, err)
 					ExpectWithOffset(1, out).To(ContainSubstring("%s was scheduled to start\n", vmName), "command \"virtctl start %s\" did not print expected message: %s", vmName, out+stderr)
 				}
